dom/tests/mochitest/pointerlock/test_mousePos.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
    <title>Test for Bug 633602</title>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
    <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
    <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>
	<p id="display"></p>
<div id="content">
	<div style="background-color:blue; width:500px; height:500px; float: left;" id="blueBox">
		<p style="color:White" id='info'>MouseLock Element</p>
	</div>
</div>
    <pre id="test">
        <script type="application/javascript">
        
            /** Test for Bug 633602 **/
            SimpleTest.waitForExplicitFinish();

			SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only", false);
            SimpleTest.waitForFocus(function() {
				var pointer = navigator.pointer;
				var blueBox = document.getElementById("blueBox");
				var xMove = window.innerWidth;
				var yMove = window.innerHeight;
				var movedX = 0;
				var movedY = 0;

				document.addEventListener("mozfullscreenchange", function() {
					if (document.mozFullScreen){
						console.log("full screen change");
						//If blueBox is the element that we have locked on to then proceed with the test
						if (document.mozFullScreenElement === blueBox) {
							pointer.lock(blueBox);
							is(pointer.islocked(), true, "Mouse should be locked when it gets a DOM element.");
							pointer.unlock();
						}
					}
				}, false);
				
				blueBox.addEventListener("mouselocklost", function(e) {
					console.log("mouselock unlocked.");
					window.blur();
					//blur the browser explicitly
					window.onblur = function() {
						console.log("firefox focus has been lost.");

						document.addEventListener("mousemove", function (e) {
							movedX = movedX + e.movementX;
							movedY = movedY + e.movementY;	
						}, false);

						//move mouse outside the browser boundaries using javascript? +50 is to be outside the boundaries of the browser
						synthesizeMouse(blueBox, xMove+50, yMove+50, { type: "mousemove" }, window);
						//then set the focus on the browser
						window.focus();

						//then check for movementX and movementY, should be 0
						if (movedX == 0){
							xtestOK = true;
						}	
						
						if (movedY == 0) {
							ytestOK = true;
						}

						ok(xtestOK, "The X Movement should be set to 0.");
						ok(ytestOK, "The Y Movement should be set to 0.");

						document.mozCancelFullScreen();
						SimpleTest.finish();
					}
				}, false);

				console.log("start test #27");
				blueBox.mozRequestFullScreen();
            });
        </script>
    </pre>
</body>
</html>
